pyproject.toml
[project]
name = "moitracrew"
version = "0.1.0"
description = "AI Development Team — CrewAI + Google Gemini"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.1.0",
    "crewai-tools>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
moitracrew = "moitracrew.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.uv]
dev-dependencies = []
